Add Moment.js as a dependency
Moment.js is added as a dependency to handle date-time functionality in the application. This library provides more accurate and efficient date-time operations. This package enables date-time manipulation in a simpler way and reduces the complexity in our codebase. Likewise, syntax arrangements in package.json were made for better organization.